server/Add show & index orphanages routes & abstract to controller
