+++ server/src/routes.ts
@@ -1,35 +1,10 @@
 import { Router } from 'express';
-import { getRepository } from 'typeorm';
-import Orphanage from './models/Orphanage';
+import OrphanagesController from './controllers/OrphanagesController'
 
 const routes = Router();
 
-routes.post('/orphanages', async (request, response) => {
-  const {
-    name,
-    latitude,
-    longitude,
-    about,
-    instructions,
-    opening_hours,
-    open_on_weekends,
-  } = request.body;
-
-  const orphanagesRepository = getRepository(Orphanage);
-
-  const orphanage = orphanagesRepository.create({
-    name,
-    latitude,
-    longitude,
-    about,
-    instructions,
-    opening_hours,
-    open_on_weekends,
-  });
-
-  await orphanagesRepository.save(orphanage);
-
-  return response.status(201).json(orphanage);
-});
+routes.get('/orphanages', OrphanagesController.index);
+routes.get('/orphanages/:id', OrphanagesController.show);
+routes.post('/orphanages', OrphanagesController.create);
 
 export default routes;
